+++ lib/handlers/404.js
@@ -1,0 +1,6 @@
+function handler(req, res){
+  if(this.debug) console.log('unsupported',req.params,req.query,req.body);
+  res.send('404 Not Supported');
+};
+
+module.exports = handler
