+++ lib/handlers/default.js
@@ -1,0 +1,6 @@
+// Example Handler
+function handler(req, res){
+        res.send(req);
+};
+
+module.exports = handler
