+++ lib/handlers/hello.js
@@ -1,0 +1,7 @@
+// Hello Handler
+
+function handler(req, res){
+  res.send('cloki');
+};
+
+module.exports = handler
